+++ MapleServer2/Packets/SkillCooldownPacket.cs
@@ -8,7 +8,7 @@ namespace MapleServer2.Packets;
 
 public static class SkillCooldownPacket
 {
-    public static PacketWriter SetCooldown(int skillId, int endTick)
+    public static PacketWriter SetCooldown(long skillId, int endTick)
     {
         PacketWriter pWriter = PacketWriter.Of(SendOp.SkillDamage);
 
@@ -21,7 +21,7 @@ public static class SkillCooldownPacket
         return pWriter;
     }
 
-    public static PacketWriter SetCooldowns(int[] skillIds, int[] endTicks)
+    public static PacketWriter SetCooldowns(long[] skillIds, int[] endTicks)
     {
         PacketWriter pWriter = PacketWriter.Of(SendOp.SkillDamage);
